skipping tests in python3.6
these tests don't pass in 3.6 or below because of
the order of kwargs given by vars(). so the tests are
incompatible because it assumes a specific order when
joining the string to build the url.

this is not a problem however because the order does not
matter in order to get data