Add feature to filter IPEA series by last n observations
IPEA API does not have a built-in API to let the user get the last N
observations, so we have to mess with what we've got, which is start and
end date.

The strategy here is to get use the SERMAXDATA fields
from the series' metadata (meaning, the maximum date), and then subtract
it N times (we have to take into account the series' periodicity here,
e.g. we need to subtract 3 quarters from a quarterly series, not three
days or months), and use this to filter every date greater than the
subtraction result.